More work on the symbol table for Fixed2Free
Due to the MOVE and MOVEL OpCodes being used for type coercion and
conversion I am having to create a symbol table. Free Format RPG has a
suite of BIFs that are used for type conversion. Therefore I need to
know the data types for the source and target fields so I can assign
the appropriate BIF